==> 0._avoin-data-suomen-tyokyvyttomyyselakkeensaajat-sairauden-mukaan.ipynb <==
# ***0._avoin-data-suomen-tyokyvyttomyyselakkeensaajat-sairauden-mukaan***
1. Suomen työkyvyttömyyseläkkeensaajat sairauden mukaan, Kelan tilastotietokanta 2003-2023.
    1. *Tiedosto*: '_avoin-data-suomen-tyokyvyttomyyselakkeensaajat-sairauden-mukaan.csv'
    1. *Verkkolähde*: https://tietotarjotin.fi/tilastodata/988121/avoin-data-suomen-tyokyvyttomyyselakkeensaajat-sairauden-mukaan

import pandas as pd

df = pd.read_csv('raw/_avoin-data-suomen-tyokyvyttomyyselakkeensaajat-sairauden-mukaan.csv') 
df

df.info()

df.isnull().any()

df.columns

df.vuosi.sort_values().unique()

df.kunta_nimi.sort_values().unique()

!***Täydennettävät tiedot:*** Koordinaatit, kunnat, kunta numerot.

***Puuttuvien arvojen prosenttiosuus:***

(df.isnull().sum()/ len(df))*100
