==> 1._wiki-Luettelo-Suomen-kunnista_clean.ipynb <==
# ***1._wiki-Luettelo-Suomen-kunnista_clean***
***Todennetaan 'wiki' aineiston sisällön riittävyys, vertaamalla 'tilastokeskuksen' aineistoon.***
1. Tilastokeskus aineisto, Kunnat 2024.
    1. *Tiedosto*: '_tilastokeskus_luokitukset-kunta_1_20240101.csv'
    1. *Verkkolähde*: https://www2.tilastokeskus.fi/fi/luokitukset/kunta/
    
    
2. Suomen aluejako aineisto, luettelo Suomen kunnista.
    1. *Tiedosto*: '_wiki-Luettelo-Suomen-kunnista.csv'
    1. *Verkkolähde:* https://fi.wikipedia.org/wiki/Luettelo_Suomen_kunnista
    


***[OUT: _wiki-Luettelo-Suomen-kunnista_clean.csv]***

import pandas as pd

dx = pd.read_csv("raw/_wiki-Luettelo-Suomen-kunnista.csv",delimiter=";")
dx = dx.replace(";",",")
dx

dx=dx.drop(columns=['Nimitys', 'Kielet','Kuntakeskus', 'Seutukunta', 'Väkiluku[5]', 'Kok.pinta-ala (km²)[6]','Maapinta-ala (km²)[6]', 'Väestötiheys(as./km²) '], axis=1)
dx

dx = dx.rename(columns={"Kunta-nro[4]":"kunta_nro","Kunnan nimi":"kunta_nimi"})
dx

dy = pd.read_csv("_tilastokeskus-luokitukset-kunta_1_20240101.csv")
dy

dy[['kunta_nro', 'kunta_nimi']] = dy['kunta_nro kunta_nimi'].str.strip().str.split(' ', n=1, expand=True)
dy.drop("kunta_nro kunta_nimi", axis=1, inplace=True)
dy

dx_kunta_nro = dx.kunta_nimi.sort_values().unique()
dy_kunta_nro = dy.kunta_nimi.sort_values().unique()

# in dx not in dy
dx_not_in_dy = set(dx_kunta_nro) - set(dy_kunta_nro)

# in dy not in dx
dy_not_in_dx = set(dy_kunta_nro) - set(dx_kunta_nro)

print("values in dx not in dy:")
print(dx_not_in_dy)

print("\nvalues in dy not in dx:")
print(dy_not_in_dx)

dx.to_csv('clean/_wiki-Luettelo-Suomen-kunnista_clean.csv', index=False)
