==> 2._wiki-Luettelo-Suomen-kuntien-koordinaateista_clean.ipynb <==
# ***2._wiki-Luettelo-Suomen-kuntien-koordinaateista_clean***

1. Suomen kuntien koordinaatit, WGS84. 1.1.1987. 
    1. *Tiedosto*: '_wiki-Luettelo-Suomen-kuntien-koordinaateista.csv'
    1. *Verkkolähde*: https://fi.wikipedia.org/wiki/Luettelo_Suomen_kuntien_koordinaateista


***[OUT: _wiki-Luettelo-Suomen-kuntien-koordinaateista_clean.csv]***

import pandas as pd 

kunnat = pd.read_csv('raw/_wiki-Luettelo-Suomen-kuntien-koordinaateista.csv', delimiter='\t')
kunnat

***Muunnetaan koordinaatit DMS muodosta DD muotoon***

def convert_to_decimal(coord):
    coord = coord.replace('°N', '').replace('°E', '').replace('°', '').replace(',', '.').strip()
    return float(coord)

kunnat['latitude'] = kunnat['Koordinaatit'].apply(lambda x: convert_to_decimal(x.split(',')[0]))
kunnat['longitude'] = kunnat['Koordinaatit'].apply(lambda x: convert_to_decimal(x.split(',')[1]))

kunnat.head()

***Huomioidaan muuttujien nimet tiedostojen yhdistämisen helpottamiseksi:***

kunnat = kunnat.rename(columns= {'Kunta':'kunta_nimi'})

***Puuttuvien arvojen prosenttiosuus:***

missing_percents = (kunnat.isnull().sum()/ len(kunnat))*100
missing_percents

***Tallennetaan uutena tiedostona:***

kunnat[['kunta_nimi', 'latitude', 'longitude']].to_csv('clean/_wiki-Luettelo-Suomen-kuntien-koordinaateista_clean.csv', index=False)
