==> 3.merge_koordinaatit&kunnat.ipynb <==
# ***3.merge1_koordinaatit&kunnat*** 
1. _wiki-Luettelo-Suomen-kunnista_clean.ipnyb 
    1. _wiki-Luettelo-Suomen-kunnista_clean.csv
1. _wiki-Luettelo-Suomen-kuntien-koordinaateista_clean.ipnyb
    1. _wiki-Luettelo-Suomen-kuntien-koordinaateista_clean.csv


***[OUT: 1.merge_koordinaatit&kunnat.csv]***

import pandas as pd 

dx = pd.read_csv("clean/_wiki-Luettelo-Suomen-kuntien-koordinaateista_clean.csv")
dy = pd.read_csv("clean/_wiki-Luettelo-Suomen-kunnista_clean.csv")

dx.kunta_nimi.nunique(), dy.kunta_nimi.nunique() 

dxdy = dy.merge(dx, on='kunta_nimi', how="left")
dxdy

***Puuttuvien arvojen prosenttiosuus:***

(dxdy.isnull().sum()/ len(dxdy))*100

dxdy.to_csv("merge/merge1_koordinaatit&kunnat.csv", index=False)
