==> 4.merge1&_avoin-data.ipynb <==
# ***4.merge1&_avoin-data_suomen_työkyvyttömyyseläkkeensaajat_2003-2023_merge***
1. _avoin-data-suomen-tyokyvyttomyyselakkeensaajat-sairauden-mukaan.csv

1. merge1_koordinaatit&kunnat.ipynb 
    1. merge1_koordinaatit&kunnat.csv

***[OUT: merge1&_avoin-data_suomen_työkyvyttömyyseläkkeensaajat_2003-2023_merge.csv]***

import warnings
warnings.filterwarnings("ignore")

import pandas as pd

dx = pd.read_csv('merge/merge1_koordinaatit&kunnat.csv', delimiter=',')
dy = pd.read_csv('raw/_avoin-data-suomen-tyokyvyttomyyselakkeensaajat-sairauden-mukaan.csv', delimiter=',')

dx['kunta_nimi'] = dx['kunta_nimi'].str.strip()
dy['kunta_nimi'] = dy['kunta_nimi'].str.strip()

dxdy = dy.merge(dx, on='kunta_nro', how="left", suffixes=('', '_right'))
dxdy

dxdy.drop(['aikatyyppi','vuosikuukausi','kunta_nimi_right'],axis=1)

new_order = ['Maakunta','kunta_nro','kunta_nimi','latitude','longitude',
             'sairauspaaryhma','ikaryhma','sukupuoli',
             'saaja_lkm','kuukausi_nro','vuosi']

dxdy = dxdy[new_order]
dxdy

***Puuttuvien arvojen prosenttiosuus:***

(dxdy.isnull().sum()/ len(dxdy))*100

***Pudotetaan rivit, joissa on puutteellista tietoa.***

dxdy.dropna(inplace=True)

(dxdy.isnull().sum()/ len(dxdy))*100

dxdy=dxdy.sort_values(by="vuosi")

dxdy.to_csv('merge/merge1&_avoin-data_suomen_työkyvyttömyyseläkkeensaajat_2003-2023_merge.csv', index=False)
